Revert "Mvtec 3d (#907)"
This reverts commit 052a3ad22f8621628e1d5c23ae151a98c40b4bca.
