+++ anomalib/data/utils/__init__.py
@@ -11,10 +11,8 @@ from .image import (
     generate_output_image_filename,
     get_image_filenames,
     get_image_height_and_width,
-    read_depth_image,
     read_image,
 )
-from .path import _check_and_convert_path, _prepare_files_labels, _resolve_path
 from .split import (
     Split,
     TestSplitMode,
@@ -31,7 +29,6 @@ __all__ = [
     "get_image_height_and_width",
     "random_2d_perlin",
     "read_image",
-    "read_depth_image",
     "random_split",
     "split_by_label",
     "concatenate_datasets",
@@ -46,7 +43,4 @@ __all__ = [
     "InputNormalizationMethod",
     "download_and_extract",
     "DownloadInfo",
-    "_check_and_convert_path",
-    "_prepare_files_labels",
-    "_resolve_path",
 ]
