+++ anomalib/data/utils/image.py
@@ -11,7 +11,6 @@ from pathlib import Path
 
 import cv2
 import numpy as np
-import tifffile as tiff
 import torch.nn.functional as F
 from torch import Tensor
 from torchvision.datasets.folder import IMG_EXTENSIONS
@@ -203,24 +202,6 @@ def read_image(path: str | Path, image_size: int | tuple[int, int] | None = None
         # prototyping new ideas.
         height, width = get_image_height_and_width(image_size)
         image = cv2.resize(image, dsize=(width, height), interpolation=cv2.INTER_AREA)
-
-    return image
-
-
-def read_depth_image(path: str | Path) -> np.ndarray:
-    """Read tiff depth image from disk.
-
-    Args:
-        path (str, Path): path to the image file
-
-    Example:
-        >>> image = read_depth_image("test_image.tiff")
-
-    Returns:
-        image as numpy array
-    """
-    path = path if isinstance(path, str) else str(path)
-    image = tiff.imread(path)
 
     return image
